Removed StoreTest & moved to ApplicationTest, get params from application.conf.test
